IMDB_Clone/requirements.txt
asgiref==3.3.4
astroid==2.5.6
colorama==0.4.4
Django==3.2.4
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
isort==5.8.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
PyJWT==2.1.0
pylint==2.8.3
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
wrapt==1.12.1
